chore: add permission guard to license page
